refactoring user
