+++ backend/modules/user/handlers/user_handler.go
@@ -1,66 +1,19 @@
-package routes
+package handlers
 
 import (
-	"backend/internal/adminpanel/entities"
-	"backend/internal/adminpanel/models"
 	"backend/internal/adminpanel/services/utils"
-	"errors"
+	"backend/modules/user/models"
+	"backend/modules/user/repository"
+	"backend/modules/user/service"
 	"fmt"
 	"github.com/gin-gonic/gin"
 	"github.com/google/uuid"
-	"gorm.io/gorm"
-	"log"
 	"net/http"
 	"strconv"
 )
 
-func LoginHandler(ctx *gin.Context) {
-	var loginRequest = entities.LoginRequest{}
-	if err := ctx.ShouldBindJSON(&loginRequest); err != nil {
-		ctx.JSON(http.StatusBadRequest, gin.H{"error": "invalid login request"})
-		return
-	}
-
-	db, ok := utils.GetDBFromContext(ctx)
-	if !ok {
-		return
-	}
-
-	user, err := models.GetUserByEmail(db, loginRequest.Email)
-	if err != nil {
-		if errors.Is(err, gorm.ErrRecordNotFound) {
-			ctx.JSON(http.StatusNotFound, gin.H{"error": "User not found"})
-			return
-		}
-		ctx.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-	if !utils.ComparePasswords(loginRequest.Password, user.Password) {
-		log.Println("Password mismatch for user:", user.Email)
-		ctx.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid credentials"})
-		return
-	}
-
-	// Tenant отримуємо з middleware-контексту
-	tenant, ok := ctx.Get("tenant")
-	if !ok {
-		ctx.JSON(http.StatusInternalServerError, gin.H{"error": "Tenant info missing"})
-		return
-	}
-
-	tenantData := tenant.(entities.Tenant)
-
-	token, err := utils.GenerateJWTToken(user.Email, user.ID, tenantData.Domain, tenantData.ID)
-	if err != nil {
-		ctx.JSON(http.StatusInternalServerError, gin.H{"error": "Could not generate token"})
-		return
-	}
-	ctx.JSON(http.StatusOK, gin.H{"access_token": token, "token_type": "bearer"})
-	log.Println("Login successful")
-}
-
 func CreateUser(ctx *gin.Context) {
-	user := new(entities.User)
+	user := new(models.User)
 	if err := ctx.ShouldBindJSON(user); err != nil {
 		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
@@ -70,7 +23,7 @@ func CreateUser(ctx *gin.Context) {
 		return
 	}
 
-	newUser, err := models.CreateUser(db, user)
+	newUser, err := repository.CreateUser(db, user)
 	if err != nil {
 		ctx.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
@@ -117,12 +70,12 @@ func ReadAllUsers(ctx *gin.Context) {
 		return
 	}
 
-	users, err := models.GetAllUsers(db, limit, skip)
+	users, err := repository.GetAllUsers(db, limit, skip)
 	if err != nil {
 		ctx.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
 	}
-	userResponses := models.TransformUsers(users)
+	userResponses := service.TransformUsers(users)
 	response := models.AllUsers{
 		Data:  userResponses,
 		Count: len(userResponses),
@@ -147,7 +100,7 @@ func UpdateCurrentUser(ctx *gin.Context) {
 		return
 	}
 
-	updatedUser, err := models.UpdateUserById(db, userID, &updateUser)
+	updatedUser, err := repository.UpdateUserById(db, userID, &updateUser)
 	if err != nil {
 		if err.Error() == "user not found" {
 			ctx.JSON(http.StatusNotFound, gin.H{"error": "User not found"})
@@ -158,36 +111,6 @@ func UpdateCurrentUser(ctx *gin.Context) {
 	}
 
 	ctx.JSON(http.StatusOK, updatedUser)
-}
-
-func UpdatePasswordCurrentUser(ctx *gin.Context) {
-	userID, ok := utils.GetUserIDFromContext(ctx)
-	if !ok {
-		return
-	}
-
-	db, ok := utils.GetDBFromContext(ctx)
-	if !ok {
-		return
-	}
-
-	var updatePassword models.UpdatePassword
-	if err := ctx.ShouldBindJSON(&updatePassword); err != nil {
-		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-	message, err := models.UpdateCurrentUserPassword(db, userID, &updatePassword)
-	if err != nil {
-		if err.Error() == "user not found" {
-			ctx.JSON(http.StatusNotFound, gin.H{"error": "User not found"})
-		} else {
-			ctx.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		}
-		return
-	}
-
-	ctx.JSON(http.StatusOK, gin.H{"message": message})
-
 }
 
 func DeleteUser(ctx *gin.Context) {
@@ -230,7 +153,7 @@ func DeleteUser(ctx *gin.Context) {
 		return
 	}
 
-	err = models.DeleteUserById(db, id)
+	err = repository.DeleteUserById(db, id)
 	if err != nil {
 		if err.Error() == "user not found" {
 			ctx.JSON(http.StatusNotFound, gin.H{"error": "User not found"})
