+++ backend/modules/user/models/user.go
@@ -1,0 +1,24 @@
+package models
+
+import (
+	"github.com/google/uuid"
+	"gorm.io/gorm"
+	"time"
+)
+
+type User struct {
+	ID          uuid.UUID `gorm:"type:uuid;primaryKey" json:"id"`
+	FullName    string    `gorm:"not null" json:"fullName"`
+	Avatar      string    `gorm:"default:null" json:"avatar"`
+	Email       string    `gorm:"unique;not null" json:"email"`
+	Password    string    `gorm:"not null" json:"password"`
+	IsActive    bool      `gorm:"default:true" json:"isActive"`
+	IsSuperUser bool      `gorm:"default:false" json:"isSuperUser"`
+	CreatedAt   time.Time
+	UpdatedAt   time.Time
+}
+
+func (user *User) BeforeCreate(*gorm.DB) error {
+	user.ID = uuid.New()
+	return nil
+}
